perf(midct): split large components and functionsto solve lint
